Relocated FalseClient for external tests to use
